Fase 1: Alembic, montos a Numeric, fixes de bugs de auditoria

+++ models.py
@@ -63,7 +63,7 @@ class Gasto(db.Model):
     __tablename__ = 'gasto'
     id = db.Column(db.Integer, primary_key=True)
     descripcion = db.Column(db.String(200), nullable=False)
-    monto = db.Column(db.Float, nullable=False)
+    monto = db.Column(db.Numeric(14, 2), nullable=False)
     fecha = db.Column(db.Date, nullable=False)
     usuario_id = db.Column(db.Integer, db.ForeignKey('usuario.id'), nullable=False)
     categoria_id = db.Column(db.Integer, db.ForeignKey('categoria.id'))
@@ -76,7 +76,7 @@ class Ingreso(db.Model):
     __tablename__ = 'ingreso'
     id = db.Column(db.Integer, primary_key=True)
     descripcion = db.Column(db.String(200), nullable=False)
-    monto = db.Column(db.Float, nullable=False)
+    monto = db.Column(db.Numeric(14, 2), nullable=False)
     fecha = db.Column(db.Date, nullable=False)
     usuario_id = db.Column(db.Integer, db.ForeignKey('usuario.id'), nullable=False)
     categoria_id = db.Column(db.Integer, db.ForeignKey('categoria.id'))
